Factor out Xcode and Git modules
