+++ lib/spm_version_updates/plugin.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 require "semantic"
-require "xcodeproj"
+require_relative "git"
+require_relative "xcode"
 
 module Danger
   # A Danger plugin for checking if there are versions upgrades available for SPM dependencies
@@ -35,18 +36,16 @@ module Danger
     # @raise [XcodeprojPathMustBeSet] if the xcodeproj_path is blank
     # @return   [void]
     def check_for_updates(xcodeproj_path)
-      raise(XcodeprojPathMustBeSet) if xcodeproj_path.nil? || xcodeproj_path.empty?
-
-      remote_packages = filter_remote_packages(Xcodeproj::Project.open(xcodeproj_path))
-      resolved_versions = get_resolved_versions(xcodeproj_path)
+      remote_packages = Xcode.get_packages(xcodeproj_path)
+      resolved_versions = Xcode.get_resolved_versions(xcodeproj_path)
       $stderr.puts("Found resolved versions for #{resolved_versions.size} packages")
 
-      self.ignore_repos = self.ignore_repos&.map! { |repo| trim_repo_url(repo) }
+      self.ignore_repos = self.ignore_repos&.map! { |repo| Git.trim_repo_url(repo) }
 
       remote_packages.each { |repository_url, requirement|
         next if self.ignore_repos&.include?(repository_url)
 
-        name = repo_name(repository_url)
+        name = Git.repo_name(repository_url)
         resolved_version = resolved_versions[repository_url]
         kind = requirement["kind"]
 
@@ -57,12 +56,12 @@ module Danger
 
         if kind == "branch"
           branch = requirement["branch"]
-          last_commit = git_branch_last_commit(repository_url, branch)
+          last_commit = Git.branch_last_commit(repository_url, branch)
           warn("Newer commit available for #{name} (#{branch}): #{last_commit}") unless last_commit == resolved_version
           next
         end
 
-        available_versions = git_versions(repository_url)
+        available_versions = Git.version_tags(repository_url)
         next if available_versions.first.to_s == resolved_version
 
         if kind == "exactVersion" && @check_when_exact
@@ -75,77 +74,6 @@ module Danger
           warn_for_new_versions_range(available_versions, name, requirement, resolved_version)
         end
       }
-    end
-
-    # Extracts resolved versions from Package.resolved relative to an Xcode project
-    # @param   [String] xcodeproj_path
-    #          The path to your Xcode project
-    # @raise [CouldNotFindResolvedFile] if no Package.resolved files were found
-    # @return [Hash<String, String>]
-    def get_resolved_versions(xcodeproj_path)
-      resolved_paths = find_packages_resolved_file(xcodeproj_path)
-      raise(CouldNotFindResolvedFile) if resolved_paths.empty?
-
-      resolved_versions = resolved_paths.map { |resolved_path|
-        JSON.load_file!(resolved_path)["pins"]
-          .to_h { |pin|
-            [trim_repo_url(pin["location"]), pin["state"]["version"] || pin["state"]["revision"]]
-          }
-      }
-      resolved_versions.reduce(:merge!)
-    end
-
-    # Extract a readable name for the repo given the url, generally org/repo
-    # @return [String]
-    def repo_name(repo_url)
-      match = repo_url.match(%r{([\w-]+/[\w-]+)(.git)?$})
-
-      if match
-        match[1] || match[0]
-      else
-        repo_url
-      end
-    end
-
-    # Find the configured SPM dependencies in the xcodeproj
-    # @param   [Xcodeproj::Project::Object] project
-    #          The Xcodeproj object of the project
-    # @return [Hash<String, Hash>]
-    def filter_remote_packages(project)
-      project.objects.select { |obj|
-        obj.kind_of?(Xcodeproj::Project::Object::XCRemoteSwiftPackageReference) &&
-          obj.requirement["kind"] != "commit"
-      }
-        .to_h { |package|
-          [trim_repo_url(package.repositoryURL), package.requirement]
-        }
-    end
-
-    # Removes protocol and trailing .git from a repo URL
-    # @param   [String] repo_url
-    #          The URL of the repository
-    # @return [String]
-    def trim_repo_url(repo_url)
-      repo_url.split("://").last.gsub(/\.git$/, "")
-    end
-
-    # Find the Packages.resolved file
-    # @return [Array<String>]
-    def find_packages_resolved_file(xcodeproj_path)
-      locations = []
-      # First check the workspace for a resolved file
-      workspace = xcodeproj_path.sub("xcodeproj", "xcworkspace")
-      if Dir.exist?(workspace)
-        path = File.join(workspace, "xcshareddata", "swiftpm", "Package.resolved")
-        locations << path if File.exist?(path)
-      end
-
-      # Then check the project for a resolved file
-      path = File.join(xcodeproj_path, "project.xcworkspace", "xcshareddata", "swiftpm", "Package.resolved")
-      locations << path if File.exist?(path)
-
-      $stderr.puts("Searching for resolved packages in: #{locations}")
-      locations
     end
 
     private
@@ -196,44 +124,5 @@ Newest version of #{name}: #{available_versions.first} (but this package is conf
         TEXT
       ) unless newest_above_reqs == newest_meeting_reqs || newest_meeting_reqs.to_s == resolved_version
     end
-
-    # Call git to list tags
-    # @param   [String] repo_url
-    #          The URL of the dependency's repository
-    # @return [Array<Semantic::Version>]
-    def git_versions(repo_url)
-      versions = `git ls-remote -t #{repo_url}`
-        .split("\n")
-        .map { |line| line.split("/tags/").last }
-        .filter_map { |line|
-          begin
-            Semantic::Version.new(line)
-          rescue ArgumentError
-            nil
-          end
-        }
-      versions.sort!
-      versions.reverse!
-      versions
-    end
-
-    # Calkl git to find the last commit on a branch
-    # @param   [String] repo_url
-    #          The URL of the dependency's repository
-    # @param   [String] branch_name
-    #          The name of the branch on which to find the last commit
-    # @return [String]
-    def git_branch_last_commit(repo_url, branch_name)
-      `git ls-remote -h #{repo_url}`
-        .split("\n")
-        .find { |line| line.split("\trefs/heads/")[1] == branch_name }
-        .split("\trefs/heads/")[0]
-    end
-  end
-
-  class XcodeprojPathMustBeSet < StandardError
-  end
-
-  class CouldNotFindResolvedFile < StandardError
   end
 end
